move bcdi.utils.constants to bcdi.constants
