+++ bcdi/postprocessing/analysis.py
@@ -29,7 +29,7 @@ import bcdi.utils.image_registration as reg
 import bcdi.utils.utilities as util
 import bcdi.utils.validation as valid
 from bcdi.experiment.setup import Setup
-from bcdi.utils.constants import AXIS_TO_ARRAY
+from bcdi.constants import AXIS_TO_ARRAY
 from bcdi.utils.text import Comment
 
 module_logger = logging.getLogger(__name__)
