+++ bcdi/postprocessing/process_scan.py
@@ -25,7 +25,7 @@ import bcdi.graph.graph_utils as gu
 import bcdi.postprocessing.postprocessing_utils as pu
 from bcdi.experiment.setup import Setup
 from bcdi.postprocessing.analysis import create_analysis
-from bcdi.utils.constants import AXIS_TO_ARRAY
+from bcdi.constants import AXIS_TO_ARRAY
 from bcdi.utils.snippets_logging import FILE_FORMATTER
 
 logger = logging.getLogger(__name__)
