+++ bcdi/postprocessing/raw_orthogonalization.py
@@ -22,7 +22,7 @@ from matplotlib import pyplot as plt
 import bcdi.utils.utilities as util
 from bcdi.experiment.setup import Setup
 from bcdi.postprocessing.analysis import create_analysis
-from bcdi.utils.constants import AXIS_TO_ARRAY
+from bcdi.constants import AXIS_TO_ARRAY
 from bcdi.utils.snippets_logging import FILE_FORMATTER
 
 logger = logging.getLogger(__name__)
