--- bcdi/utils/constants.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# BCDI: tools for pre(post)-processing Bragg coherent X-ray diffraction imaging data
-#   (c) 07/2017-06/2019 : CNRS UMR 7344 IM2NP
-#   (c) 07/2019-06/2022 : DESY PHOTON SCIENCE
-#       authors:
-#         Jerome Carnis, carnis_jerome@yahoo.fr
-
-"""Definition of global constants."""
-
-import numpy as np
-
-AXIS_TO_ARRAY = {
-    "x": np.array([1, 0, 0]),
-    "y": np.array([0, 1, 0]),
-    "z": np.array([0, 0, 1]),
-}  # in xyz order
-
-LOGGING_LEVELS = ("debug", "info", "warning", "error", "exception")
